refactor(app): use app layout component to wrap pages structure
