+++ src/pages/admin/trilhas.tsx
@@ -1,10 +1,9 @@
 import { orangeEvolutionLogo } from 'components/@constants'
+import { AppLayout } from 'components/AppLayout'
 import { ButtonLink } from 'components/ButtonLink'
-import { Header } from 'components/Header'
 import { Heading } from 'components/Heading'
 import { Text } from 'components/Text'
 import { TrailForm } from 'components/TrailForm'
-import { useTheme } from 'contexts/ThemeContext'
 import { getCookie } from 'cookies-next'
 import useWindowSize from 'hooks/useWindowSize'
 import 'keen-slider/keen-slider.min.css'
@@ -16,7 +15,6 @@ import { User } from 'libs/user/types'
 import { GetServerSideProps } from 'next'
 import Head from 'next/head'
 import Image from 'next/image'
-import { BodyWrapper, Main } from 'styles/pages/home'
 import {
   ButtonWrapper,
   Card,
@@ -48,8 +46,6 @@ export default function Trails({ token, user }: TrailsProps) {
     }
   })
 
-  const { theme } = useTheme()
-
   const { width } = useWindowSize()
 
   const { isLoading } = useTrails(token)
@@ -61,80 +57,77 @@ export default function Trails({ token, user }: TrailsProps) {
       <Head>
         <title>Orange Evolution - Trilhas</title>
       </Head>
-      <BodyWrapper theme={theme}>
-        <Header />
-        <Main style={{ padding: '2rem 0' }}>
-          <Image
-            src={orangeEvolutionLogo}
-            alt="Logo da orange evolution"
-            width={197}
-            height={122}
-          />
-          <Heading asChild size="lg">
-            <h1>Escolha sua trilha!</h1>
-          </Heading>
 
-          <TextWrapper>
-            <Text asChild>
-              <p>
-                O Orange Evolution consiste em trilhas totalmente gratuitas para
-                que você possa iniciar a sua carreira na tecnologia. Você terá
-                acesso a vídeos, lives, artigos, apostilas e até cursos
-                gratuitos, além desses conteúdos serem da Orange Juice, de
-                parceiros e empresas que confiamos.
-              </p>
-            </Text>
+      <AppLayout>
+        <Image
+          src={orangeEvolutionLogo}
+          alt="Logo da orange evolution"
+          width={197}
+          height={122}
+        />
+        <Heading asChild size="lg">
+          <h1>Escolha sua trilha!</h1>
+        </Heading>
 
-            <br />
+        <TextWrapper>
+          <Text asChild>
+            <p>
+              O Orange Evolution consiste em trilhas totalmente gratuitas para
+              que você possa iniciar a sua carreira na tecnologia. Você terá
+              acesso a vídeos, lives, artigos, apostilas e até cursos gratuitos,
+              além desses conteúdos serem da Orange Juice, de parceiros e
+              empresas que confiamos.
+            </p>
+          </Text>
 
-            <Text asChild>
-              <p>
-                Essa trilha foi montada pensando em quem está começando na área,
-                ou passando por uma migração de carreira e ainda não sabe
-                exatamente o que é esse mundo. Então, aperta o cinto e vem com a
-                gente nessa jornada!
-              </p>
-            </Text>
-          </TextWrapper>
+          <br />
 
-          <TrailForm />
+          <Text asChild>
+            <p>
+              Essa trilha foi montada pensando em quem está começando na área,
+              ou passando por uma migração de carreira e ainda não sabe
+              exatamente o que é esse mundo. Então, aperta o cinto e vem com a
+              gente nessa jornada!
+            </p>
+          </Text>
+        </TextWrapper>
 
-          <CardListWrapper>
-            <CardList
-              ref={width <= 768 ? sliderRef : null}
-              className={width <= 768 ? 'keen-slider' : ''}
-            >
-              {trails.map((trail) => (
-                <Card
-                  key={trail.id}
-                  theme={theme}
-                  className={width <= 768 ? 'keen-slider__slide' : ''}
-                  css={{
-                    paddingTop: '2.5rem'
-                  }}
-                >
-                  <Heading size="sm">{trail.title}</Heading>
+        <TrailForm />
 
-                  <TrailForm trail={trail} />
+        <CardListWrapper>
+          <CardList
+            ref={width <= 768 ? sliderRef : null}
+            className={width <= 768 ? 'keen-slider' : ''}
+          >
+            {trails.map((trail) => (
+              <Card
+                key={trail.id}
+                className={width <= 768 ? 'keen-slider__slide' : ''}
+                css={{
+                  paddingTop: '2rem'
+                }}
+              >
+                <Heading size="sm">{trail.title}</Heading>
 
-                  <CardImage
-                    src={trail.icon_url}
-                    alt={trail.title}
-                    width={100}
-                    height={100}
-                  />
+                <TrailForm trail={trail} />
 
-                  <ButtonWrapper>
-                    <ButtonLink href={`/admin/trilha/${trail.id}`}>
-                      Acesssar
-                    </ButtonLink>
-                  </ButtonWrapper>
-                </Card>
-              ))}
-            </CardList>
-          </CardListWrapper>
-        </Main>
-      </BodyWrapper>
+                <CardImage
+                  src={trail.icon_url}
+                  alt={trail.title}
+                  width={100}
+                  height={100}
+                />
+
+                <ButtonWrapper>
+                  <ButtonLink href={`/admin/trilha/${trail.id}`}>
+                    Acesssar
+                  </ButtonLink>
+                </ButtonWrapper>
+              </Card>
+            ))}
+          </CardList>
+        </CardListWrapper>
+      </AppLayout>
     </>
   )
 }
