+++ src/pages/trilhas.tsx
@@ -1,6 +1,6 @@
 import { orangeEvolutionLogo } from 'components/@constants'
+import { AppLayout } from 'components/AppLayout'
 import { ButtonLink } from 'components/ButtonLink'
-import { Header } from 'components/Header'
 import { Heading } from 'components/Heading'
 import { Progress } from 'components/Progress'
 import { Text } from 'components/Text'
@@ -16,7 +16,6 @@ import { User } from 'libs/user/types'
 import { GetServerSideProps } from 'next'
 import Head from 'next/head'
 import Image from 'next/image'
-import { BodyWrapper, Main } from 'styles/pages/home'
 import {
   ButtonWrapper,
   Card,
@@ -61,82 +60,77 @@ export default function Trails({ token, user }: TrailsProps) {
       <Head>
         <title>Orange Evolution - Trilhas</title>
       </Head>
-      <BodyWrapper theme={theme}>
-        <Header />
-        <Main style={{ padding: '2rem 0' }}>
-          <Image
-            src={orangeEvolutionLogo}
-            alt="Logo da orange evolution"
-            width={197}
-            height={122}
-          />
-          <Heading asChild size="lg">
-            <h1>Escolha sua trilha!</h1>
-          </Heading>
+      <AppLayout>
+        <Image
+          src={orangeEvolutionLogo}
+          alt="Logo da orange evolution"
+          width={197}
+          height={122}
+        />
+        <Heading asChild size="lg">
+          <h1>Escolha sua trilha!</h1>
+        </Heading>
 
-          <TextWrapper>
-            <Text asChild>
-              <p>
-                O Orange Evolution consiste em trilhas totalmente gratuitas para
-                que você possa iniciar a sua carreira na tecnologia. Você terá
-                acesso a vídeos, lives, artigos, apostilas e até cursos
-                gratuitos, além desses conteúdos serem da Orange Juice, de
-                parceiros e empresas que confiamos.
-              </p>
-            </Text>
+        <TextWrapper>
+          <Text asChild>
+            <p>
+              O Orange Evolution consiste em trilhas totalmente gratuitas para
+              que você possa iniciar a sua carreira na tecnologia. Você terá
+              acesso a vídeos, lives, artigos, apostilas e até cursos gratuitos,
+              além desses conteúdos serem da Orange Juice, de parceiros e
+              empresas que confiamos.
+            </p>
+          </Text>
 
-            <br />
+          <br />
 
-            <Text asChild>
-              <p>
-                Essa trilha foi montada pensando em quem está começando na área,
-                ou passando por uma migração de carreira e ainda não sabe
-                exatamente o que é esse mundo. Então, aperta o cinto e vem com a
-                gente nessa jornada!
-              </p>
-            </Text>
-          </TextWrapper>
+          <Text asChild>
+            <p>
+              Essa trilha foi montada pensando em quem está começando na área,
+              ou passando por uma migração de carreira e ainda não sabe
+              exatamente o que é esse mundo. Então, aperta o cinto e vem com a
+              gente nessa jornada!
+            </p>
+          </Text>
+        </TextWrapper>
 
-          <CardListWrapper>
-            <CardList
-              ref={width <= 768 ? sliderRef : null}
-              className={width <= 768 ? 'keen-slider' : ''}
-            >
-              {trails.map((trail) => (
-                <Card
-                  key={trail.id}
-                  theme={theme}
-                  className={width <= 768 ? 'keen-slider__slide' : ''}
-                >
-                  <Heading size="sm">{trail.title}</Heading>
+        <CardListWrapper>
+          <CardList
+            ref={width <= 768 ? sliderRef : null}
+            className={width <= 768 ? 'keen-slider' : ''}
+          >
+            {trails.map((trail) => (
+              <Card
+                key={trail.id}
+                theme={theme}
+                className={width <= 768 ? 'keen-slider__slide' : ''}
+              >
+                <Heading size="sm">{trail.title}</Heading>
 
-                  <CardImage
-                    src={trail.icon_url}
-                    alt={trail.title}
-                    width={100}
-                    height={100}
-                  />
+                <CardImage
+                  src={trail.icon_url}
+                  alt={trail.title}
+                  width={100}
+                  height={100}
+                />
 
-                  <Progress
-                    isTrailPage
-                    donePercentage={
-                      trail.total === 0
-                        ? 0
-                        : (trail.completed / trail.total) * 100
-                    }
-                  />
+                <Progress
+                  isTrailPage
+                  donePercentage={
+                    trail.total === 0
+                      ? 0
+                      : (trail.completed / trail.total) * 100
+                  }
+                />
 
-                  <ButtonWrapper>
-                    <ButtonLink href={`/trilha/${trail.id}`}>
-                      Acesssar
-                    </ButtonLink>
-                  </ButtonWrapper>
-                </Card>
-              ))}
-            </CardList>
-          </CardListWrapper>
-        </Main>
-      </BodyWrapper>
+                <ButtonWrapper>
+                  <ButtonLink href={`/trilha/${trail.id}`}>Acesssar</ButtonLink>
+                </ButtonWrapper>
+              </Card>
+            ))}
+          </CardList>
+        </CardListWrapper>
+      </AppLayout>
     </>
   )
 }
